submodules/PhysX-Anything/requirements.txt
absl-py==2.2.2
academictorrents==2.3.3
accelerate==1.4.0
addict==2.4.0
aiohttp==3.9.3
aiosignal==1.3.1
albumentations==1.4.3
aliyun-python-sdk-core==2.15.1
aliyun-python-sdk-kms==2.16.3
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asttokens==3.0.0
async-timeout==4.0.3
attrs==23.2.0
av==11.0.0
bencode.py==2.0.0
bitstring==3.1.5
blessed==1.20.0
blinker==1.7.0
blobfile==2.0.2
boto3 ==1.34.2
botocore==1.34.59
braceexpand==0.1.7
cachetools==5.5.2
ccimport==0.4.2
certifi==2022.12.7
cffi==1.16.0
chamferdist==1.0.3
charset-normalizer==2.0.12
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.2.0
crcmod==1.7
cryptography==43.0.0
cumm==0.4.11
cumm-cu118==0.4.11
cycler==0.12.1
Cython==3.0.10
dataclasses-json==0.6.4
decorator==5.2.1
Deprecated==1.2.14
diffusers==0.20.2
docker-pycreds==0.4.0
easydict==1.13
einops==0.7.0
emd==0.0.0
enlighten==1.12.4
entrypoints==0.4
environs==10.0.0
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.14.0
fire==0.6.0
flash-attn==2.5.8
Flask==3.0.2
flatbuffers==23.5.26
fonttools==4.45.1
freetype-py==2.4.0
frozenlist==1.4.1
fsspec==2023.10.0
ftfy==6.2.0
future==0.16.0
gitdb==4.0.11
GitPython==3.1.42
glcontext==3.0.0
glfw==2.7.0
google-auth==2.38.0
gpytoolbox==0.2.0
grpcio==1.71.0
h5py==3.10.0
hatchling==1.25.0
hf-xet==1.2.0
huggingface-hub==0.36.0
humanfriendly==10.0
humanize==4.9.0
idna==3.4
igraph==0.11.8
imageio==2.34.0
imageio-ffmpeg==0.4.7
imhist==0.0.4
importlib_metadata==8.0.0
imWatermark==0.0.2
invisible-watermark==0.2.0
ipdb==0.13.13
ipycanvas==0.13.1
ipyevents==2.0.2
ipython==8.27.0
ipywidgets==8.1.2
itsdangerous==2.1.2
jaxtyping==0.2.31
jedi==0.19.2
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.4.0
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==7.4.9
jupyter_core==5.7.2
jupyterlab_widgets==3.0.10
kaolin==0.15.0
kiwisolver==1.4.5
kornia==0.6.0
lark==1.1.9
lazy_loader==0.3
lightning-utilities==0.11.2
llvmlite==0.42.0
lpips==0.1.4
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.2
matplotlib-inline==0.1.7
mdurl==0.1.2
memory-profiler==0.61.0
mesh-to-sdf==0.0.15
meshio==5.3.4
mmengine==0.10.4
model-index==0.1.11
moderngl==5.12.0
mpmath==1.3.0
multidict==6.0.5
multiprocessing-logging==0.3.4
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.0
neuralnet-pytorch==0.0.3
ninja==1.11.1.1
numba==0.59.0
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.7.0.84
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.19.3
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
onnxruntime==1.16.3
open-clip-torch==2.7.0
openai-clip==1.0.1
opencv-python==4.8.1.78
opencv-python-headless==4.9.0.80
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.1
ordered-set==4.1.0
oss2==2.17.0
packaging==24.1
pandas==2.2.2
parso==0.8.4
pathspec==0.12.1
pccm==0.4.11
petrel-oss-sdk===v2.2.1-2-g1505ef3-master
pexpect==4.9.0
Pillow==10.1.0
platformdirs==4.2.0
pluggy==1.5.0
plyfile==1.0.2
pointnet2-cuda==0.0.0
pooch==1.8.1
portalocker==2.8.2
prefixed==0.7.1
prompt_toolkit==3.0.50
protobuf==5.27.3
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybind11==2.12.0
pycolmap==0.6.1
pycparser==2.22
pycryptodome==3.20.0
pyglet==1.5.28
pygltflib==1.16.2
Pygments==2.19.1
PyMatting==1.1.12
PyMCubes==0.1.4
pymeshfix==0.17.0
PyOpenGL==3.1.0
PyOpenGL-accelerate==3.1.7
pyparsing==3.1.1
PyPubSub==3.3.0
pyrender==0.1.45
pysdf==0.1.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytorch-lightning==2.1.2
pytorch-pretrained-vit==0.0.7
pytz==2023.4
pyvista==0.44.2
PyWavelets==1.6.0
PyYAML==6.0.1
pyzmq==24.0.1
qwen-vl-utils==0.0.14
referencing==0.33.0
regex==2023.12.25
rembg==2.0.55
requests==2.28.2
rich==13.4.2
rpds-py==0.18.0
rsa==4.9
rtree==1.4.0
s3transfer==0.9.0
safetensors==0.4.5
scikit-image==0.22.0
scikit-learn==1.4.2
scipy==1.11.4
scooby==0.10.0
sentry-sdk==1.40.6
setproctitle==1.3.3
six==1.16.0
slangtorch==1.2.2
smmap==5.0.1
spconv-cu118==2.3.6
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
taming-transformers==0.0.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
texttable==1.7.0
thop==0.1.1.post2209072238
threadpoolctl==3.4.0
tifffile==2023.9.26
timm==0.9.10
tokenizers==0.21.4
tomli==2.2.1
torch==2.1.1+cu118
torch-scatter==2.1.2+pt21cu118
torchaudio==2.1.1+cu118
torchmcubes==0.1.0
torchmetrics==1.3.2
torchvision==0.16.1+cu118
tornado==6.4
tqdm==4.65.2
traitlets==5.14.3
transformers==4.50.0
transforms3d==0.4.2
trimesh==4.0.5
triton==2.1.0
trove-classifiers==2024.7.2
typeguard==2.13.3
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==1.26.18
usd-core==23.5
utils3d @ git+https://github.com/EasternJournalist/utils3d.git@9a4eb15e4021b67b12c460c7057d642626897ec8
visdom==0.2.4
vtk==9.3.1
wandb==0.17.7
wcwidth==0.2.13
webdataset==0.2.86
websocket-client==1.8.0
Werkzeug==3.1.3
widgetsnbextension==4.0.10
wrapt==1.16.0
xatlas==0.0.9
xformers==0.0.23+cu118
yapf==0.40.2
yarl==1.9.4
zipp==3.19.2
